Replaced selenium code with new method in main script.
